P2P Logging is movef to the separate module
